add direct payment fee method
