+++ packages/cryptoplease/lib/core/payments/direct_payments/direct_payment_fee_method.dart
@@ -1,0 +1,8 @@
+/// Method for the fee payment in direct transfer feature.
+enum DirectPaymentFeeMethod {
+  /// Transaction created using solana lib directly.
+  v1,
+
+  /// Transaction created using cryptoplease API to pay the fees in USDC.
+  v2,
+}
