+++ packages/cryptoplease/lib/features/incoming_split_key_payment/bl/bloc.dart
@@ -1,8 +1,8 @@
 import 'package:bloc_concurrency/bloc_concurrency.dart';
 import 'package:cryptoplease/core/processing_state.dart';
-import 'package:cryptoplease/core/split_key_payments/split_key_api_version.dart';
-import 'package:cryptoplease/core/split_key_payments/transaction/tx_creator.dart';
-import 'package:cryptoplease/core/split_key_payments/transaction/tx_creator_strategy.dart';
+import 'package:cryptoplease/core/payments/split_key_payments/split_key_api_version.dart';
+import 'package:cryptoplease/core/payments/split_key_payments/transaction/tx_creator.dart';
+import 'package:cryptoplease/core/payments/tx_creator_strategy.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/bl/models.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/bl/repository.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/bl/tx_processor.dart';
