+++ packages/cryptoplease/lib/features/request_pay/bl/request_pay_bloc.dart
@@ -2,8 +2,9 @@ import 'package:cryptoplease/config.dart';
 import 'package:cryptoplease/core/accounts/bl/account.dart';
 import 'package:cryptoplease/core/amount.dart';
 import 'package:cryptoplease/core/currency.dart';
+import 'package:cryptoplease/core/payments/direct_payments/direct_payment_fee_method.dart';
+import 'package:cryptoplease/core/payments/split_key_payments/split_key_api_version.dart';
 import 'package:cryptoplease/core/processing_state.dart';
-import 'package:cryptoplease/core/split_key_payments/split_key_api_version.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/create_outgoing_transfer_bloc/ft/bloc.dart';
 import 'package:cryptoplease/features/outgoing_transfer/bl/outgoing_payment.dart';
@@ -88,7 +89,7 @@ class RequestPayBloc extends Bloc<_Event, _State> {
         amount: amount,
         tokenAddress: state.amount.token.address,
         tokenType: OutgoingTransferTokenType.fungibleToken,
-        apiVersion: SplitKeyApiVersion.v2,
+        feeMethod: DirectPaymentFeeMethod.v2,
         state: OutgoingTransferState.draft(
           encodedTx: directPayment.transaction,
         ),
