+++ packages/cryptoplease/test/features/incoming_split_key_payment/bloc_test.dart
@@ -1,5 +1,5 @@
-import 'package:cryptoplease/core/split_key_payments/split_key_api_version.dart';
-import 'package:cryptoplease/core/split_key_payments/transaction/tx_creator_strategy.dart';
+import 'package:cryptoplease/core/payments/split_key_payments/split_key_api_version.dart';
+import 'package:cryptoplease/core/payments/tx_creator_strategy.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/core/wallet.dart';
 import 'package:cryptoplease/features/incoming_split_key_payment/bl/bloc.dart';
